feat(voucher): Implement new voucher schema and migration script
- Added detailed data dictionary for voucher tables, including attributes and business rules.
- Created new SQL migration script to restructure the voucher schema for marketplace workflow.
- Preserved legacy tables for rollback/reference and migrated base data from old voucher structure to new schema.
- Established foreign key relationships and constraints in the new schema.
- Included checks for data integrity and consistency during migration.

+++ Marketplace-platform/src/main/java/docker_test/com/mappers/product/ProductMapper.java
@@ -35,10 +35,10 @@ public final class ProductMapper implements IMapper<Product> {
             product.setReview_count(rs.getInt(StringValue.PRODUCT_REVIEW_COUNT_COL));
             
             // Kích thước / Cân nặng
-            product.setWeight(rs.getDouble(StringValue.PRODUCT_WEIGHT_COL));
-            product.setLength(rs.getDouble(StringValue.PRODUCT_LENGTH_COL));
-            product.setWidth(rs.getDouble(StringValue.PRODUCT_WIDTH_COL));
-            product.setHeight(rs.getDouble(StringValue.PRODUCT_HEIGHT_COL));
+            product.setWeight(rs.getInt(StringValue.PRODUCT_WEIGHT_COL));
+            product.setLength(rs.getInt(StringValue.PRODUCT_LENGTH_COL));
+            product.setWidth(rs.getInt(StringValue.PRODUCT_WIDTH_COL));
+            product.setHeight(rs.getInt(StringValue.PRODUCT_HEIGHT_COL));
             
             product.setBrand(rs.getString(StringValue.PRODUCT_BRAND_COL));
             product.setIs_active(rs.getInt(StringValue.PRODUCT_ACTIVE_COL));
